updated api directory structure
