+++ api/Dockerfile
@@ -1,16 +1,14 @@
-FROM python:3.9 as base
+FROM python:3.10-slim AS base
 
 WORKDIR /app
 
 COPY requirements.txt .
-RUN pip install --no-cache-dir -r requirements.txt
+RUN pip install -r requirements.txt
+
+FROM base AS builder
 
 COPY . .
 
-ENV PYTHONPATH=/app:$PYTHONPATH
+FROM builder as service
 
-FROM base as service
-
-WORKDIR /app/src
-
-CMD ["uvicorn", "main:app", "--host", "0.0.0.0", "--port", "8080"]
+CMD ["python", "main.py"]
