+++ api/app/helpers/api_router.py
@@ -1,0 +1,23 @@
+from typing import Any, Callable
+
+from fastapi import APIRouter as FastAPIRouter
+from fastapi.types import DecoratedCallable
+
+
+class APIRouter(FastAPIRouter):
+    def api_route(
+        self, path: str, *, include_in_schema: bool = True, **kwargs: Any
+    ) -> Callable[[DecoratedCallable], DecoratedCallable]:
+        if path.endswith("/"):
+            path = path[:-1]
+
+        add_path = super().api_route(path, include_in_schema=include_in_schema, **kwargs)
+
+        alternate_path = path + "/"
+        add_alternate_path = super().api_route(alternate_path, include_in_schema=False, **kwargs)
+
+        def decorator(func: DecoratedCallable) -> DecoratedCallable:
+            add_alternate_path(func)
+            return add_path(func)
+
+        return decorator
