+++ api/app/models/responses/optimize.py
@@ -1,0 +1,3 @@
+from typing import List
+
+OptimizeResponse = List[List[dict]]
