+++ api/requirements.txt
@@ -9,3 +9,4 @@ bs4
 pulp
 requests
 lxml
+orjson
